refactor: move resource-assertions
