+++ packages/opentelemetry-resources/src/resource-assertions.ts
@@ -15,7 +15,7 @@
  */
 
 import * as assert from 'assert';
-import { Resource } from '../../src/Resource';
+import { Resource } from './Resource';
 import {
   CLOUD_RESOURCE,
   CONTAINER_RESOURCE,
@@ -23,7 +23,7 @@ import {
   K8S_RESOURCE,
   TELEMETRY_SDK_RESOURCE,
   SERVICE_RESOURCE,
-} from '../../src/constants';
+} from './constants';
 /**
  * Test utility method to validate a cloud resource
  *
